renames module sensor to sensors
